backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=2.0.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
